+++ app/src/test/java/com/github/barriosnahuel/vossosunboton/feature/vault/ImmersiveListenScreenTest.kt
@@ -51,6 +51,7 @@ internal class ImmersiveListenScreenTest : AbstractRobolectricTest() {
         AnalyticsTrackerProvider.setForTest(fakeTracker)
         mockkObject(PlayerControllerFactory)
         every { PlayerControllerFactory.instance.setOnStartStopListener(any()) } answers { nothing }
+        every { PlayerControllerFactory.instance.removeOnStartStopListener(any()) } answers { nothing }
     }
 
     @After